refactored ProgressDialog
