+++ src/components/dialogs/DialogContainer.vue
@@ -4,6 +4,7 @@
   <IdentifierDialog />
   <ExportDialog />
   <ErrorDialog />
+  <ProgressDialog />
 </template>
 <script setup lang="ts">
 import DeleteDialog from "./DeleteDialog.vue";
@@ -11,4 +12,5 @@ import ErrorDialog from "./ErrorDialog.vue";
 import ExportDialog from "./ExportDialog.vue";
 import IdentifierDialog from "./IdentifierDialog.vue";
 import ImportDialog from "./ImportDialog.vue";
+import ProgressDialog from "./ProgressDialog.vue";
 </script>
