+++ src/components/dialogs/ProgressDialog.vue
@@ -1,6 +1,6 @@
 <template>
   <q-dialog
-    v-model="visible"
+    v-model="progressDialog.visible"
     persistent
   >
     <q-card style="width: 50vw">
@@ -10,31 +10,31 @@
       <q-card-section class="q-pt-none">
         <q-linear-progress
           size="1rem"
-          :value="progress"
+          :value="progressDialog.progress"
           animation-speed="0"
         >
           <div class="absolute-full flex flex-center">
             <q-badge color="transparent">
-              {{ Math.round(progress * 100) }} %
+              {{ Math.round(progressDialog.progress * 100) }} %
             </q-badge>
           </div>
         </q-linear-progress>
         <div
-          v-for="(error, index) in errors"
+          v-for="(error, index) in progressDialog.errors"
           :key="index"
           class="text-negative"
           :data-cy="`error-prompt-${index}`"
         >
           {{ error }}
         </div>
-        <div v-show="progress >= 0.999">
+        <div v-show="progressDialog.progress >= 0.999">
           {{ $t("files-were-successfully-transfer-to-new-path") }}
         </div>
       </q-card-section>
       <q-card-actions align="right">
         <q-btn
           flat
-          :disable="progress < 0.999"
+          :disable="progressDialog.progress < 0.999"
           :ripple="false"
           label="OK"
           v-close-popup
@@ -46,20 +46,5 @@
   </q-dialog>
 </template>
 <script setup lang="ts">
-import { computed, PropType } from "vue";
-
-const props = defineProps({
-  modelValue: { type: Boolean, required: true },
-  progress: { type: Number, required: true },
-  errors: Object as PropType<Error[]>,
-});
-const emit = defineEmits(["update:modelValue"]);
-const visible = computed({
-  get() {
-    return props.modelValue;
-  },
-  set(visibility: boolean) {
-    emit("update:modelValue", visibility);
-  },
-});
+import { progressDialog } from "./dialogController";
 </script>
